Do not realize all configurations automatically
Fixes #28192
Fixes #15639
Partially fixes #26418 (We will not deprecate auto building, just change behavior and warn about it in upgrade guide. Detecting this is hard, and avoiding this behavior relies on buildscript changes that we don't want to propogate eveywhere)

Deprecate DefaultArtifactPublicationSet (used by KMP, they will have to change their plugin logic to have the assemble task directly depend on their artifact).
This removes implicit behavior where any artifact registered with this set is 'intelligently' selected to be built by the assemble task.
Now, the jar, war, and ear are all built when running assemble

Add artifacts directly to the archives configuration in the war, ear, java, and distributions plugins.
In 9.0, we will add a dependency directly from the assmeble task to these artifacts, skipping the archives configuration.
For now, we don't want to change behavior if users expect these artifacts to be in the archives configuration.

Deprecate the archvies configuration.

Deprecate adding artifacts to the archives configuration (You can still add archives to non-consumable configurations and therefore configurations that are deprecated for consumption, as Android and Kotlin use non-consumable configurations for publishing)

Add a test to verify that we do not automatically realize registered configurations in WellBehavedPluginTest
Add tests to verify new deprecations
Add test to verify assemble task can be referenced directly with dependsOn to add artifacts to be built during that lifecycle task