Do not realize unrealted configurations in target project when consuming variants
Fixes #28192

We avoid realizing lazy configurations in the target project that are known to be non-consumable.
We leverage the public type of the pending Provider, as that encodes the known role of the configuration.

For unlocked configurations without a role, we still need to realize the configuration to determine if it is consumable or not.

This may show improvements in some projects that heavily use lazily registered configurations.

Come 9.0, Gradle's Java plugins will also register role-locked lazy configurations. We will likely see improvements then.

This change is to open the pathway for ecosystem plugins to create more Configurations without fear of the overhead associated with them.